Add better View and Model handling
This change adds better View and Model handling.

First off, we introduce the `Observable` class, which exposes on function
`.on(eventName, callback)` which fires if the Object itself has changes.
This in and of itself seems like a good idea, but after extensive
testing, isn't a so good idea. Some changes are needed.

Second, we introduce the `View` class, which exposes a `render` and
`remove` function common and overridable for all `View`s. The meaning of
this is to create a common API for View stuff so we can be sure that the
element always is created etc. No error handling is present for this, so
if you pass in an empty element, well then, shit happens. Needs work.

All in all, I'm quite pleased with having touched Object.observe,
template string and EventEmitter in one pull request. Hopefully more to
come.